Add missing linkage declaration to Cargo.toml
Per rust documentation (https://doc.rust-lang.org/cargo/reference/build-scripts.html#the-links-manifest-key),
a crate which links with a native C library is supposed to declare that
using the "links" manifest key in order to avoid linking two copies of
a native library. Some third-party sites also use this field to aid in
searching for rust bindings to native libraries.