Switch to Rust pkg-config library
Rust has a pkg-config library that is maintained by the Rust foundation.
Using this instead of the hand-written integration reduces the effort
required to maintain the build system and also adds support for
cross-compilation (something which is important to many people and
organizations in the DPDK community). This library also handles a few
edge-cases I ran into with the previous approach involving custom
configurations of DPDK which link libbsd (a secondary -L argument), or
statically link mlx5_core into librte_ethdev.